refactor : CarRacingGame에서 생성한 기능을 GameSetting으로 이동
- 신규 클래스 GameSetting 생성
- CarRacingGame.racing에서 Scanner로 자동차 대수, 시도 횟수를 받던 것을 GameSetting 클래스로 이동
- Test 코드도 GameSettingTest로 이동
