+++ README.md
@@ -2,7 +2,7 @@
 
 ## 기능 목록
 - [ ] 게임에 참여하는 자동차 수와 이동 횟수를 입력받는다.
-- [ ] 랜덤한 0-9 정수 중 4 이상이 나오면 전진, 그렇지 않다면 제자리에 머무른다.
+- [ ] 현재 위치를 포함하는 n개의 자동차를 만들고  랜덤한 0-9 정수 중 4 이상이 나오면 전진, 그렇지 않다면 제자리에 머무른다.
 - [ ] 현재 자동차들의 상태를 화면에 출력한다.
 
 ## Commit message 구분
